Module ftoken::ftoken_mod::state

Structs

AuctionInfo

BidInfo

bid information as stored by ftoken contract

PropInfo

proposal information as stored by ftoken contract

PropInfoTally

ResvVote

Reservation price and stake stored in binary (serialized U192) representing a Uint128 with additional 19 decimal points

StakedTokens

U192

Little-endian large integer type

U256

Little-endian large integer type

U384

Little-endian large integer type

VoteRegister

Enums

Vote

Vote cast on proposals

VoteResult

Constants

AGGRESVPRICE_STORE

ALLOWED_TOKENS

AUCTION_INFO

CURRENT_PROP_ID

FTKN_CONFIG

FTKN_STAKE

FTOKEN_CONTR_FTKN

NFT_VIEW_KEY

PREFIX_BIDS

PREFIX_UNDR_NFT

PROPS_STORE

RESVPRICE_STORE

VOTES_BUCKET

VOTES_TOTAL

Functions

add_bid

agg_resv_price_r

agg_resv_price_w

Aggregate reservation price

auction_info_r

auction_info_w

information on auction

ftkn_config_r

ftkn_config_w

config specifically for ftoken functionality

ftkn_stake_r

ftkn_stake_w

staked ftokens

ftoken_info_r

ftoken_info_w

FtokenContr storage: stores information on this ftokens contract

get_bid

get_bids

Returns all bids in reverse order and the number of bids. If bids can be queried, bidders’ addressed can be identified (even if addresses are removed from the query result, if using side chain attacks)

get_last_bid

may_get_bid_from_addr

nft_vk_r

nft_vk_w

stores viewing key to query nft contract

prop_id_r

prop_id_w

index the next proposal to be received

props_r

props_w

Proposal storage: stores proposal information

resv_price_r

resv_price_w

Reservation price votes for each address

set_bid

votes_r

votes_total_r

votes_total_w

proposal vote tally, which is the running cumulative tally

votes_w

Multilevel bucket to store proposal votes. Key intended to be [bid_id, HumanAddr]